move files and clean up import functions
