+++ lair_of_xaldern/player_manager.py
